requirements.txt
Django==1.7
argparse==1.2.1
django-polymorphic==0.5.6
factory-boy==2.4.1
ipython==2.2.0
wsgiref==0.1.2

# Common Code / Entropy
git+https://github.com/commoncode/entropy.git

# Common Code / Economica
git+https://github.com/economica/django-rea.git